requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.1
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
lxml==4.9.3

# For development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 